# ARROW-11064: [Rust][DataFusion] Speed up hash join on smaller batches