FYI @andygrove @jorgecarleitao 

This is related to findings here: https://github.com/apache/arrow/pull/9036
This is a ~20% win on smaller batch sizes, but also has an effect on bigger sizes (about 10% on size of 16k locally)

Query 12 with batch size 4096

PR:

```
Query 12 iteration 0 took 484.0 ms
Query 12 iteration 1 took 482.5 ms
Query 12 iteration 2 took 496.9 ms
Query 12 iteration 3 took 488.1 ms
Query 12 iteration 4 took 504.9 ms
Query 12 iteration 5 took 490.4 ms
Query 12 iteration 6 took 486.8 ms
Query 12 iteration 7 took 499.3 ms
Query 12 iteration 8 took 485.0 ms
Query 12 iteration 9 took 488.8 ms
Query 12 avg time: 490.67 ms

```

Master:
```
Query 12 iteration 0 took 669.6 ms
Query 12 iteration 1 took 764.8 ms
Query 12 iteration 2 took 705.8 ms
Query 12 iteration 3 took 645.1 ms
Query 12 iteration 4 took 640.9 ms
Query 12 iteration 5 took 639.2 ms
Query 12 iteration 6 took 658.9 ms
Query 12 iteration 7 took 643.2 ms
Query 12 iteration 8 took 639.0 ms
Query 12 iteration 9 took 639.2 ms
Query 12 avg time: 664.57 ms
```