@jorgecarleitao 2020-12-29
Thanks a lot @Dandandan , how these pisky `clone`s are affecting performance :(