@Dandandan 2020-12-30
@jorgecarleitao Yes, small things can have a big effect...

The difference is a bit "magnified" though because of the `n*n` behavior in the join.
If you have ideas of how to solve this issue, please share.

What looks like is the problem here:

* for the left side of the join we  `.iter()` on the left side batches iteration (a couple of times, in the join implementation itself and in `MutableDataArray`) when we process a probe-side batch.
* if we halve the batch size, the number of items in the build-side array grows by two, but also the number of probe-side batches we process in the join, thus the `n*n` slowdown.